Makes Drawable-trait depend on Transformable, moves draw code to trait
This should make it trivial to make new objects drawable.
